adjusted can frame periodicity for example networks due to changes in FiCo4OMNeT

+++ src/gateway/buffering/AccumulationGatewayBuffering.cc
@@ -173,7 +173,7 @@ void AccumulationGatewayBuffering::emitSignals(cMessagePointerList* poolList) {
         emit(poolHoldUpTimeSignals[poolID], holdUpTime);
     }
     arrivalTimes->clear();
-    emit(poolSizeSignals[poolID], poolList->size());
+    emit(poolSizeSignals[poolID], static_cast<unsigned long> (poolList->size()));
 }
 
 unsigned int AccumulationGatewayBuffering::getPoolID(
